Add submission handling and prepare for the v0.3.0 release
Add handling of submissions to subreddits. To achieve this, there are now
two distinct classes for all the logic pertaining to submissions and
comments. On start, the bot forks. The child process takes care of
submissions, while the parent process takes care of comments. Prepare
for the v0.3.0 release of the bot by bumping the version number in
the sample configuration file. Update the documentation and changelog
accordingly.